requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-instrumentation-fastapi>=0.40b0
opentelemetry-exporter-prometheus>=1.12.0rc1
opentelemetry-instrumentation>=0.40b0
prometheus-client>=0.17.0
aioredis>=2.0.1
httpx>=0.24.0
python-json-logger>=2.0.7